Adding insertions, CMake configuration, and tests.
Files changed location to src/ and test/, to better get along with
CMake.
Added an integrity check function check(). It tests some, but not
all, of the properties a B-tree must have. That is, it is
necessary, but not sufficient for correctness.
Added insertion into the tree, along with some tests for it.