Why test hundreds when we can test... millions?
Changed a test to insert millions of elements instead of
hundreds, checking for absence before, and presence after
insertion.

The modulus is chosen to be a prime p, such that p > 2*n,
which guarantees that there will be no duplicates in the list
[x^2 % p | x <- [0..n]]. If a^2 = b^2 (mod p), then
p | (a + b)(a - b), which cannot happen if both are at most p/2.